Move InputForm from molecules to organisms
